python/markitdown_service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.3.0
python-multipart==0.0.9
pytesseract==0.3.10
markitdown==0.0.1a3
